+++ inbox/views.py
@@ -1,17 +1,23 @@
+import base64
+
+from django.contrib.auth import get_user_model
+from django.core.signing import Signer
 from django.utils import timezone
 from rest_framework import status
 from rest_framework.decorators import action
-from rest_framework.exceptions import ValidationError
+from rest_framework.exceptions import ValidationError, NotFound, PermissionDenied
 from rest_framework.mixins import RetrieveModelMixin, DestroyModelMixin, ListModelMixin, UpdateModelMixin
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet
 from rest_framework_extensions.mixins import NestedViewSetMixin
 
-from inbox.models import Message
+from inbox.models import Message, MessagePreferences
 from inbox.permissions import IsOwner
 from inbox.serializers import MessageSerializer, MessageListSerializer, MessageUpdateSerializer
 from inbox.utils import save_message_preferences
+
+User = get_user_model()
 
 
 class NestedMessagePreferencesMixin:
@@ -124,3 +130,82 @@ class MessageViewSet(RetrieveModelMixin, UpdateModelMixin, DestroyModelMixin, Ge
         serializer = MessageSerializer(instance, context=self.get_serializer_context())
 
         return Response(serializer.data)
+
+
+class MessagePreferencesViewSet(RetrieveModelMixin, UpdateModelMixin, GenericViewSet):
+    queryset = MessagePreferences.objects.all()
+
+    def retrieve(self, request, pk, *args, **kwargs):
+        signer = Signer()
+        signature = base64.urlsafe_b64decode(pk)
+        user_id = signer.unsign(signature.decode('utf-8'))
+
+        # Check if authenticated, make sure matches the token sent
+        if request.user.is_authenticated and user_id != str(request.user.pk):
+            raise PermissionDenied
+
+        try:
+            user = User.objects.get(pk=user_id)
+        except User.DoesNotExist:
+            return NotFound()
+
+        return Response({'results': user.message_preferences.groups}, status=status.HTTP_200_OK)
+
+    def update(self, request, pk, *args, **kwargs):
+        signer = Signer()
+        signature = base64.urlsafe_b64decode(pk)
+        user_id = signer.unsign(signature.decode('utf-8'))
+
+        # Check if authenticated, make sure matches the token sent
+        if request.user.is_authenticated and user_id != str(request.user.pk):
+            raise PermissionDenied
+
+        message_preferences = User.objects.get(pk=user_id).message_preferences
+
+        try:
+            message_preferences = save_message_preferences(message_preferences, self.request.data.get('results'))
+        except ValueError as e:
+            raise ValidationError(e.args)
+
+        return Response({'results': message_preferences.groups}, status=status.HTTP_200_OK)
+
+    @action(methods=['GET', 'PUT'], detail=True, url_path='(?P<preference_id>[a-z_]+)/(?P<medium_id>[a-z_]+)')
+    def message_preferences(self, request, pk, preference_id, medium_id, *args, **kwargs):
+        """
+        Supports updating a single preference+medium combo, eg PUT /api/v1/message-preferences/{token}/{preferenceId}/{medium}
+
+        :param request:
+        :param pk:
+        :param preference_id:
+        :param medium_id:
+        :param args:
+        :param kwargs:
+        :return:
+        """
+        signer = Signer()
+        signature = base64.urlsafe_b64decode(pk)
+        user_id = signer.unsign(signature.decode('utf-8'))
+
+        # Check if authenticated, make sure matches the token sent
+        if request.user.is_authenticated and user_id != str(request.user.pk):
+            raise PermissionDenied
+
+        message_preferences = User.objects.get(pk=user_id).message_preferences
+
+        if self.request.method == 'PUT':
+            try:
+                save_message_preferences(
+                    message_preferences,
+                    self.request.data,
+                    preference_id,
+                    medium_id
+                )
+            except ValueError as e:
+                raise ValidationError(e.args)
+            d = self.request.data
+        elif self.request.method == 'GET':
+            for group in message_preferences.groups:
+                if group['id'] == preference_id:
+                    d = group[medium_id]
+
+        return Response(d, status=status.HTTP_200_OK)